pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watermarker"
version = "0.1.0"
description = "Python CLI for batch image watermarking with logo or text"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "watermarker maintainers" }
]
keywords = ["watermark", "images", "cli", "batch", "pillow", "typer"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Graphics"
]
dependencies = [
  "Pillow>=10.0.0,<12.0.0",
  "rich>=13.7.0,<14.0.0",
  "typer>=0.12.3,<1.0.0"
]

[project.urls]
Homepage = "https://github.com/example/add-watermark"
Repository = "https://github.com/example/add-watermark"
Issues = "https://github.com/example/add-watermark/issues"

[project.scripts]
watermarker = "watermarker.cli:main"
wm = "watermarker.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

